requirements.txt
Flask
bravado-core
attrs